v0.3.1143 follow-on — SCALE-SEAM ㉞ ㉟ ㊱, structure / RFI / logistics
Move eleven methods out of client.ts into existing mixins (no extra ApiClient wrapper, no version bump — tag lag already at the bound):

- ㉞ does this frame stand? → model.ts (analytical summary, gravity solve, OpenSees, Code_Aster, lateral)
- ㉟ what does this model still need? → topics.ts (readiness, promote-to-BCF, NL-QA)
- ㊱ what resources are on site when? → schedule.ts (logistics get/put/state)

Pin 2,440 → 2,323. MARKS widened to ㊲.
