+++ apps/web/src/api/surface.test.ts
@@ -151,6 +151,9 @@ describe("the API client's public surface", () => {
       "appraisal", "listingAutofill", "shareListing", "syndicateListing",
       "schedule4d", "sequenceClash", "scheduleCpm", "evm",           // 4D + sequence clash + earned value
       "pullPlanBoard", "pullPlanMetrics", "leanPpc", "pullPlanStream",
+      "analyticalSummary", "structureSolve", "structureLateral",
+      "rfiReadiness", "rfiReadinessBcf", "rfiQa",
+      "getLogistics", "putLogistics", "logisticsState",
       "clashMetrics", "clashClearanceRules", "clashMatrix",
       "chartOfAccounts", "namingConventions", "elementsByDiscipline", "pipelineFunnel",
       "twoSidedBudget", "mepPressureLoss", "mepTrayFill", "mepThermalLoads",
