+++ apps/web/src/shell/roadmapLanes.test.ts
@@ -72,7 +72,9 @@ const OPEN = LINES.slice(0, GATED_AT === -1 ? LINES.length : GATED_AT);
 // cannot spell is a code it cannot contradict. Doing it late is how that happened; doing it in
 // the shipping release is how it stops happening.
 // Widened to ㉞ on 2026-09-01, in the release that SHIPS ㉝. Same rule.
-const MARKS = "①②③④⑤⑥⑦⑧⑨⑩⑪⑫⑬⑭⑮⑯⑰⑱⑲⑳㉑㉒㉓㉔㉕㉖㉗㉘㉙㉚㉛㉜㉝㉞";
+// Widened to ㊲ on 2026-09-01, in the follow-on that SHIPS ㉞ ㉟ ㊱ without a version bump
+// (tag lag already at the test_release_current bound). Same rule: widen on the way OUT.
+const MARKS = "①②③④⑤⑥⑦⑧⑨⑩⑪⑫⑬⑭⑮⑯⑰⑱⑲⑳㉑㉒㉓㉔㉕㉖㉗㉘㉙㉚㉛㉜㉝㉞㉟㊱㊲";
 
 /**
  * One source for the marker vocabulary, because there were **two** and they had already drifted.
